requirements.txt
GPUtil
SimpleITK
accelerate
albumentations
black
cryptography
einops
fairscale
flake8
h5py
hydra-core
hydra-optuna-sweeper
indexed-gzip  # Faster decompression in Nibabel.
ipywidgets
iterative-stratification
joblib
jupyterlab
matplotlib
mkl
wandb
monai==1.0.*
nibabel
numpy
omegaconf
opencv-python-headless
openpyxl  # Read `.xlsx` files in Pandas.
pandas
pydicom
pylint
pymongo==3.10.1
pytest
openmim
rich
scikit-image
scikit-learn
scipy
seaborn
tensorboard
torchinfo
tqdm
setuptools==59.5.0
timm

# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning==1.9.1
torchmetrics==0.11.0

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
pyarrow
fastparquet